Moved core functionality from script to buildcal library

+++ charis/primitives/offset_cal.py
@@ -62,7 +62,8 @@ def calc_offset(psflets, image, offsets, dx=64,
     outarr = np.zeros((psflets.shape[0], ny, nx))
 
     for i in range(0, nx, dx):
-        corrvals_all = matutils.crosscorr(psflets, image.data, image.ivar,
+        corrvals_all = matutils.crosscorr(psflets, image.data.astype('float64'),
+                                          image.ivar.astype('float64'),
                                           offsets, maxproc=maxcpus,
                                           m1=i, m2=i + dx)
